🐛 fix(bom): 修复需求计算列名与装柜量类型不匹配
- calculate_bom_demand 误用批次表列名 qty_on_hand 查汇总库存表 inventory，改为 quantity
- container_qty 数据库列为 INTEGER，Rust 侧 Option<i64> 解码报错，统一改为 Option<i32>
- 同步修复 get_material_by_id 解码同一列的同款隐患

+++ src-tauri/src/commands/material.rs
@@ -415,7 +415,7 @@ pub struct SaveMaterialParams {
     pub net_weight_kg: Option<f64>,
     pub gross_weight_kg: Option<f64>,
     pub packing_method: Option<String>,
-    pub container_qty: Option<i64>,
+    pub container_qty: Option<i32>,
 }
 
 /// 生成下一个物料编码（格式由系统配置决定，默认 `M-0001`）
